Adding required files
Added the resource files needed to run the project. See README for more
detailed instructions as to how to set things up. I suspect we'll have
to get together to set up each programming environment, since it was a
pain in the ass for me at least. Maybe the project settings will
magically work...